Refactoring Active Record Implementation
