+++ lib/active_record/connection_adapter.rb
